Fix dependency bug when using multiple transforms which depend on each others output
Turn the YAML nodes for each task into a sequence of nodes with the
previous task as the single element of a mapping.
Previously, all tasks were members of one mapping, thus the order of
evaluating the tasks depended on the `dict` implementation used
by the python interpreter instead of order of occurrence in the recipe.