extractors: rewrite column data conversion to convert only explicitly given columns
Instead of using a heuristic to determine whether a column
could usefully be converted to a categorical data type, convert only
columns explicitly named in the `categorical_columns` parameter to a
(automatically derived) categorical data type. Columns given in the
`numerical_columns` parameter will be converted to `float`, or to the
type given explicitly in a dictionary.