extractors: don't use default parameter values for arguments that take mutable objects Default parameter values are evaluated at function definition (see https://docs.python.org/3/reference/compound_stmts.html#function-definitions).
If one calls e.g. a constructor twice with the parameter value not set
and assigns it to an instance attribute, both instances will share the
same object. If one of the instances mutates its attribute, it will be
visible in the other instance. Thus, it is better to not use default
parameter values in a constructor. Or don't use them with anything other
than None or immutable values.