DataAttributes: adds common root to attributes
the common root is used when reading many files from different folders
matching a regex. Files with equal names were overwritten when exported
as only the input filename was used to derive the output file name.
Now, the common root can be used to determine a relative file path for
all input files and is used by the FileResultProcessor to create the
same folder structure for feather files, preventing files from being
overwritten.