performance: prevent unnecessary constructiions of format strings
building format strings containing dataframes for logging is costly.
These strings are built for every log level, even when the current
level is above. In cases that are executed frequently (extractors,
transformers, exporters), it is cheaper to first check whether
buildling the format string makes sense and skip it otherwise.