run_recipe: add the ability to run yamllint over a recipe
This is a very rudimentary implementation and just runs yamllint in
a subprocess for a nicely colored and formatted output.

Right now this outputs a lot of indentation errors in the example
recipes, none of which are of consequence to the correctness of the
resulting data produced by the framework.